Day 20: use a formula, not a map lookup, to get coordinate IDs
Given a pair `(a, b)` we can turn that into a single int with a
formula like `a + b * m`, where `m` is some magic number.  To avoid
clashes, we need this m to be such that there are no solutions other
than the trivial to:

    a + b * m == c + d * m

    a \in Z
    b \in Z
    m \in Z

This has solutions of the form:

    c = a + m * n
    d = b - n

    n \in Z

Fortunately we have constraints on our variables.  We know the length
of the input, `l`, and we know that the range of the variables `a`,
`b`, `c`, and `d` is `[-l..l]`, if the entire input consisted of going
in a straight line.

If `m = 3l`, we get the trivial solution `n = 0` (the two coordinates
are the same), but no nontrivial solutions because `n` is constrained
to be an integer:

    c = a + 3l * n
    d = b - n

    -l <= a <= l
    -l <= b <= l
    -l <= c <= l
    -l <= d <= l
    n \in Z

The maximum value of `a` (`l`) and the largest negative value of
`n` (`-1`) gives `c = l - 3l`, which is below the minimum value that
`c` can be.  Similarly for the minimum value of `a` and the smallest
positive value of `n`.

So `m = 3l` is a safe magic number.