d23p2: replace broken sphere/cube intersection algorithm (needs benchmark)
Use https://gist.github.com/myndzi/19d434b1ffbe98be89b40255e3202e98

The idea is that the cost of getting to the cube on one axis is the
cost of moving to up to the lower face (if you're below it) or down to
the higher face (if you're above it), or 0 if you're inside it.  As
Manhattan distance works by just summing the axis-distances, we can
work out these costs for each axis and just add them together.

This generalises nicely to an arbitrary number of dimensions.